simplified delay()

+++ examples/shared/fsbl/delay.h
@@ -7,9 +7,9 @@ static inline void udelay(uint32_t usec)
 	// So if usec=2 = 1.10us, and usec=10 = 2.87us,
 	// then 8 loops = 48cycles is 1.77us, so 1 cycle = 36.87ns so clock is at 27MHz (assuming SYSRAM has no
 	// wait states)
-	const unsigned long loops_per_us = 5;
-	const unsigned long loops = (unsigned long)usec * loops_per_us;
-	volatile unsigned long t = loops;
+	const uint32_t loops_per_us = 5;
+	const uint32_t loops = usec * loops_per_us;
+	volatile uint32_t t = loops;
 	while (t)
 		t = t - 1;
 }
